Implement StreamRelay cog for Discord
This code implements a StreamRelay cog for Discord that relays messages to a web client and allows control of streams via commands.

You may not like the code, the code may not work for you, and thats ok. I personally needed the code for a VERY specific use case (cable tv on discord) and yea.